requirements.txt
# requirements.txt
openai
colorama
tiktoken
playwright
python-dotenv